# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Military Store API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface LoginRequest {
  username: string;
  password: string;
}

export interface AdminInfo {
  id: number;
  username: string;
  isActive: boolean;
}

export interface LoginResponse {
  success: boolean;
  admin: AdminInfo;
}

export interface MessageResponse {
  message: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Product {
  id: number;
  name: string;
  description: string;
  price: number;
  imageUrl?: string | null;
  categoryId?: number | null;
  categoryName?: string | null;
  stock: number;
  featured: boolean;
  isActive: boolean;
  createdAt: string;
}

export interface CreateProductRequest {
  name: string;
  description: string;
  price: number;
  imageUrl?: string | null;
  categoryId?: number | null;
  stock: number;
  featured?: boolean;
  isActive?: boolean;
}

export interface UpdateProductRequest {
  name?: string;
  description?: string;
  price?: number;
  imageUrl?: string | null;
  categoryId?: number | null;
  stock?: number;
  featured?: boolean;
  isActive?: boolean;
}

export interface Category {
  id: number;
  name: string;
  description?: string | null;
  slug: string;
  imageUrl?: string | null;
  productCount: number;
}

export interface CreateCategoryRequest {
  name: string;
  description?: string | null;
  slug: string;
  imageUrl?: string | null;
}

export interface StoreSettings {
  id: number;
  storeName: string;
  storeDescription?: string | null;
  whatsappNumber: string;
  heroTitle: string;
  heroSubtitle?: string | null;
  heroImageUrl?: string | null;
  logoUrl?: string | null;
  address?: string | null;
  email?: string | null;
  operatingHours?: string | null;
}

export interface UpdateSettingsRequest {
  storeName?: string;
  storeDescription?: string | null;
  whatsappNumber?: string;
  heroTitle?: string;
  heroSubtitle?: string | null;
  heroImageUrl?: string | null;
  logoUrl?: string | null;
  address?: string | null;
  email?: string | null;
  operatingHours?: string | null;
}

export type GetProductsParams = {
  categoryId?: number;
  search?: string;
  featured?: boolean;
};