# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Military Store API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication operations
  - name: products
    description: Product operations
  - name: categories
    description: Category operations
  - name: settings
    description: Store settings operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/login:
    post:
      operationId: adminLogin
      tags: [auth]
      summary: Login as admin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/logout:
    post:
      operationId: adminLogout
      tags: [auth]
      summary: Logout current session
      responses:
        "200":
          description: Logged out successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  /auth/me:
    get:
      operationId: getAuthMe
      tags: [auth]
      summary: Get current authenticated admin
      responses:
        "200":
          description: Current admin info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminInfo"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /products:
    get:
      operationId: getProducts
      tags: [products]
      summary: Get all products (public)
      parameters:
        - name: categoryId
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
        - name: featured
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a new product (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductRequest"
      responses:
        "201":
          description: Product created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"

  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a product by ID (public)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Product not found
    put:
      operationId: updateProduct
      tags: [products]
      summary: Update a product (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProductRequest"
      responses:
        "200":
          description: Product updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Product deleted

  /categories:
    get:
      operationId: getCategories
      tags: [categories]
      summary: Get all categories (public)
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      summary: Create a new category (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryRequest"
      responses:
        "201":
          description: Category created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"

  /categories/{id}:
    put:
      operationId: updateCategory
      tags: [categories]
      summary: Update a category (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryRequest"
      responses:
        "200":
          description: Category updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
    delete:
      operationId: deleteCategory
      tags: [categories]
      summary: Delete a category (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Category deleted

  /settings:
    get:
      operationId: getSettings
      tags: [settings]
      summary: Get store settings (public)
      responses:
        "200":
          description: Store settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StoreSettings"
    put:
      operationId: updateSettings
      tags: [settings]
      summary: Update store settings (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSettingsRequest"
      responses:
        "200":
          description: Settings updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StoreSettings"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]

    LoginRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required: [username, password]

    LoginResponse:
      type: object
      properties:
        success:
          type: boolean
        admin:
          $ref: "#/components/schemas/AdminInfo"
      required: [success, admin]

    AdminInfo:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        isActive:
          type: boolean
      required: [id, username, isActive]

    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]

    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        price:
          type: number
        imageUrl:
          type: string
          nullable: true
        categoryId:
          type: integer
          nullable: true
        categoryName:
          type: string
          nullable: true
        stock:
          type: integer
        featured:
          type: boolean
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, name, description, price, stock, featured, isActive, createdAt]

    CreateProductRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        imageUrl:
          type: string
          nullable: true
        categoryId:
          type: integer
          nullable: true
        stock:
          type: integer
        featured:
          type: boolean
        isActive:
          type: boolean
      required: [name, description, price, stock]

    UpdateProductRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        imageUrl:
          type: string
          nullable: true
        categoryId:
          type: integer
          nullable: true
        stock:
          type: integer
        featured:
          type: boolean
        isActive:
          type: boolean

    Category:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        slug:
          type: string
        imageUrl:
          type: string
          nullable: true
        productCount:
          type: integer
      required: [id, name, slug, productCount]

    CreateCategoryRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        slug:
          type: string
        imageUrl:
          type: string
          nullable: true
      required: [name, slug]

    StoreSettings:
      type: object
      properties:
        id:
          type: integer
        storeName:
          type: string
        storeDescription:
          type: string
          nullable: true
        whatsappNumber:
          type: string
        heroTitle:
          type: string
        heroSubtitle:
          type: string
          nullable: true
        heroImageUrl:
          type: string
          nullable: true
        logoUrl:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        operatingHours:
          type: string
          nullable: true
      required: [id, storeName, whatsappNumber, heroTitle]

    UpdateSettingsRequest:
      type: object
      properties:
        storeName:
          type: string
        storeDescription:
          type: string
          nullable: true
        whatsappNumber:
          type: string
        heroTitle:
          type: string
        heroSubtitle:
          type: string
          nullable: true
        heroImageUrl:
          type: string
          nullable: true
        logoUrl:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        operatingHours:
          type: string
          nullable: true