# lib/api-zod/src/generated/types/adminInfo.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Military Store API
 * OpenAPI spec version: 0.1.0
 */

export interface AdminInfo {
  id: number;
  username: string;
  isActive: boolean;
}