# lib/api-zod/src/generated/types/loginResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Military Store API
 * OpenAPI spec version: 0.1.0
 */
import type { AdminInfo } from "./adminInfo";

export interface LoginResponse {
  success: boolean;
  admin: AdminInfo;
}